Fix drag and drop moving and copying files
Whilst the fix is not necessarily a good one, if the dropEffect for a given event is none (in local testing it always is) then the dispatchEvent method will dispatch a "none" tool event, which is not handled. Therefore the explorer enters a state in which it is not working properly and the file has not been moved properly. Relying on the last one triggered event seems to work just fine, there the dropEffect is move or copy.